+++ src/main/java/org/cotato/csquiz/common/config/JdaConfig.java
@@ -3,7 +3,7 @@ package org.cotato.csquiz.common.config;
 import lombok.RequiredArgsConstructor;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
-import org.cotato.csquiz.domain.education.service.DiscordButtonListener;
+import org.cotato.csquiz.domain.education.service.component.DiscordButtonListener;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
